requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
cffi==1.14.3
colorama==0.4.3
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.1.2
django-colorfield==0.3.2
django-cors-headers==3.5.0
django-extensions==3.0.9
django-filter==2.4.0
django-heroku==0.3.1
django-rest-passwordreset==1.1.0
django-sslserver==0.22
djangorestframework==3.12.1
gunicorn==20.0.4
isort==5.5.4
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
Pillow==8.0.1
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==19.1.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
